Refactor API views into organized modules
Split the large apps/api/views.py file into focused modules:
- experiments.py: Experiment-related views
- sessions.py: Experiment session views
- participants.py: Participant data management views
- files.py: File download views
- channels.py: Channel and messaging views

This improves maintainability and code organization while
preserving all existing functionality and imports.
